// src/store/chat/action.types.ts
export enum ChatActionTypes {
  CLEAR_CHAT = 'CHAT:CLEAR_CHAT',
  ADD_USER_MESSAGE = 'CHAT:ADD_USER_MESSAGE',
  ADD_EXAMPLE = 'CHAT:ADD_EXAMPLE',
  ASK_OPEN_AI = 'CHAT:ASK_OPEN_AI',
  ASK_OPEN_AI_SUCCESS = 'CHAT:ASK_OPEN_AI_SUCCESS',
  UPDATE_SYSTEM_MESSAGE = 'CHAT:UPDATE_SYSTEM_MESSAGE',
  UPDATE_GPT_MESSAGE = 'CHAT:UPDATE_GPT_MESSAGE',
  SET_MESSAGES = 'CHAT:SET_MESSAGES',
}